test cases a vymazanie socketov
