+++ simulator final/index.html
@@ -1,0 +1,154 @@
+<!DOCTYPE html>
+<html lang="sk">
+<head>
+  <meta charset="UTF-8">
+  <title>📋 Zoznam mojich úloh</title>
+  <style>
+    body {
+      font-family: 'Segoe UI', sans-serif;
+      background: #f0f0f0;
+      padding: 2rem;
+      max-width: 600px;
+      margin: auto;
+    }
+
+    h1 {
+      text-align: center;
+      color: #333;
+    }
+
+    ul {
+      list-style: none;
+      padding: 0;
+      margin-top: 1rem;
+    }
+
+    li {
+      background: #fff;
+      margin: 0.5rem 0;
+      padding: 0.75rem;
+      border-radius: 8px;
+      box-shadow: 0 1px 4px rgba(0,0,0,0.1);
+      display: flex;
+      flex-direction: column;
+    }
+
+    .description, .due-date {
+      font-size: 0.9rem;
+      color: #666;
+      margin-top: 0.3rem;
+    }
+
+    .completed {
+      background-color: #d4edda;
+    }
+
+    .to-delete {
+      background-color: #f8d7da;
+      text-decoration: line-through;
+    }
+
+    #loginSection {
+      text-align: center;
+      margin-bottom: 2rem;
+    }
+  </style>
+</head>
+<body>
+
+<h1>📋 Zoznam mojich úloh</h1>
+
+<!-- Login sekcia -->
+<div id="loginSection">
+  <input type="text" id="menoPouzivatela" placeholder="Zadaj svoje meno *" required>
+  <br>
+  <button onclick="prihlasitSa()">Prihlásiť sa</button>
+</div>
+
+<!-- Sekcia zoznamu úloh -->
+<ul id="ulohyZoznam" style="display: none;"></ul>
+
+<script>
+  let aktualnyPouzivatel = null;
+
+  function prihlasitSa() {
+    const menoInput = document.getElementById("menoPouzivatela");
+    const meno = menoInput.value.trim();
+    if (!meno) {
+      alert("Prosím, zadaj svoje meno!");
+      return;
+    }
+    aktualnyPouzivatel = meno;
+    document.getElementById("loginSection").style.display = "none";
+    document.getElementById("ulohyZoznam").style.display = "block";
+    nacitajUlohy();
+  }
+
+  async function nacitajUlohy() {
+    if (!aktualnyPouzivatel) return;
+    try {
+      const res = await fetch(`http://127.0.0.1:5000/ulohy?user_name=${encodeURIComponent(aktualnyPouzivatel)}`);
+      const data = await res.json();
+      renderUlohy(data);
+    } catch (err) {
+      console.error("Chyba pri načítaní úloh:", err);
+    }
+  }
+
+  function renderUlohy(data) {
+    const zoznam = document.getElementById("ulohyZoznam");
+    zoznam.innerHTML = "";
+    data.forEach(uloha => {
+      const li = document.createElement("li");
+      if (uloha.done) li.classList.add("completed");
+
+      const text = document.createElement("div");
+      text.textContent = uloha.text;
+      li.appendChild(text);
+
+      if (uloha.description) {
+        const desc = document.createElement("div");
+        desc.textContent = uloha.description;
+        desc.className = "description";
+        li.appendChild(desc);
+      }
+
+      if (uloha.due_date) {
+        const due = document.createElement("div");
+        due.textContent = `Do: ${uloha.due_date}`;
+        due.className = "due-date";
+        li.appendChild(due);
+      }
+
+      const button = document.createElement("button");
+      button.textContent = uloha.done ? "↩️ Zrušiť splnenie" : "✅ Splniť";
+      button.onclick = () => toggleDone(uloha);
+      li.appendChild(button);
+
+      zoznam.appendChild(li);
+    });
+  }
+
+  async function toggleDone(uloha) {
+    try {
+      await fetch(`http://127.0.0.1:5000/ulohy/${uloha.id}`, {
+        method: "PUT",
+        headers: { "Content-Type": "application/json" },
+        body: JSON.stringify({
+          text: uloha.text,
+          description: uloha.description,
+          due_date: uloha.due_date,
+          done: !uloha.done
+        })
+      });
+      nacitajUlohy();
+    } catch (err) {
+      console.error("Chyba pri aktualizácii úlohy:", err);
+    }
+  }
+
+</script>
+
+</body>
+</html>
+
